refactored actual hal logic to main component
